+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - **Customizable verbose names** for draft sharing, implemented in a non-breaking way. (@stevejalim)
 - **Customizable position** for the draft sharing menu item. (@stevejalim)
 - **GitHub Actions CI** using `tox` and Python versions 3.9 to 3.12. (@stevejalim)
-
+- Add fallback handling if the Clipboard API is not available. (@mixxorz)
 
 ### Changed